+++ circumflex-core/src/main/scala/exception.scala
@@ -8,3 +8,9 @@ class CircumflexException(msg: String, cause: Throwable = null)
     extends RuntimeException(msg, cause) {
   def this(cause: Throwable) = this(null, cause)
 }
+
+class ValidationException(val errors: Seq[Msg])
+    extends CircumflexException("Validation failed.") {
+  def this(msg: Msg) = this(List(msg))
+  def this(msg: Msg, msgs: Msg*) = this(List(msg) ++ msgs.toSeq)
+}