+++ circumflex-orm/src/test/scala/specs.scala
@@ -3,6 +3,7 @@ package ru.circumflex.orm
 import org.specs.runner.JUnit4
 import org.specs.Specification
 import xml._
+import ru.circumflex.core._
 
 object Sample {
   def createSchema = new DDLUnit(Country, City, Capital, Developer, Project, Membership).CREATE